requirements.txt
2gis==1.3.1
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
djangorestframework==3.8.2
idna==2.7
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
